num_rec/requirements.txt
tensorflow
flask
gunicorn
pillow
numpy